+++ bundled-libs/pear/http_request2/HTTP/Request2/Adapter/Mock.php
@@ -13,7 +13,7 @@
  * @category  HTTP
  * @package   HTTP_Request2
  * @author    Alexey Borzov <avb@php.net>
- * @copyright 2008-2022 Alexey Borzov <avb@php.net>
+ * @copyright 2008-2023 Alexey Borzov <avb@php.net>
  * @license   http://opensource.org/licenses/BSD-3-Clause BSD 3-Clause License
  * @link      http://pear.php.net/package/HTTP_Request2
  */
@@ -52,7 +52,7 @@ class HTTP_Request2_Adapter_Mock extends HTTP_Request2_Adapter
     /**
      * A queue of responses to be returned by sendRequest()
      *
-     * @var array
+     * @var array<int, array{HTTP_Request2_Response|Exception, ?string}>
      */
     protected $responses = [];
 
@@ -80,18 +80,17 @@ class HTTP_Request2_Adapter_Mock extends HTTP_Request2_Adapter
                 break;
             }
         }
-        if (!$response) {
-            return self::createResponseFromString("HTTP/1.1 400 Bad Request\r\n\r\n");
 
-        } elseif ($response instanceof HTTP_Request2_Response) {
+        if ($response instanceof HTTP_Request2_Response) {
             return $response;
-
-        } else {
+        } elseif ($response instanceof Exception) {
             // rethrow the exception
             $class   = get_class($response);
             $message = $response->getMessage();
             $code    = $response->getCode();
             throw new $class($message, $code);
+        } else {
+            return self::createResponseFromString("HTTP/1.1 400 Bad Request\r\n\r\n");
         }
     }