+++ bundled-libs/pear/net_dns2/Net/DNS2/Header.php
@@ -63,7 +63,7 @@ class Net_DNS2_Header
     /**
      * Constructor - builds a new Net_DNS2_Header object
      *
-     * @param mixed &$packet either a Net_DNS2_Packet object or null
+     * @param Net_DNS2_Packet &$packet either a Net_DNS2_Packet object or null
      *
      * @throws Net_DNS2_Exception
      * @access public